requirements-telegram.txt
# requirements for telegram bot
python-telegram-bot==20.8
requests
python-dotenv